# backend/docs/openapi-exportai.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ExportAI API",
    "description": "API do ExportAI para recomendacao personalizada de mercados internacionais por NCM ou HS6.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/jornadas": {
      "post": {
        "tags": [
          "Jornadas an�nimas"
        ],
        "summary": "Criar Jornada",
        "operationId": "criar_jornada_api_v1_jornadas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarJornadaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CriarJornadaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jornadas/{journey_id}": {
      "get": {
        "tags": [
          "Jornadas an�nimas"
        ],
        "summary": "Obter Jornada",
        "operationId": "obter_jornada_api_v1_jornadas__journey_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JornadaPublica"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jornadas/{journey_id}/mercados": {
      "put": {
        "tags": [
          "Jornadas an�nimas"
        ],
        "summary": "Selecionar Mercados",
        "operationId": "selecionar_mercados_api_v1_jornadas__journey_id__mercados_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelecionarMercadosRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JornadaPublica"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jornadas/{journey_id}/diagnostico-payload": {
      "get": {
        "tags": [
          "Jornadas an�nimas"
        ],
        "summary": "Preparar Diagnostico",
        "operationId": "preparar_diagnostico_api_v1_jornadas__journey_id__diagnostico_payload_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepararDiagnosticoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jornadas/{journey_id}/diagnostico": {
      "put": {
        "tags": [
          "Jornadas an�nimas"
        ],
        "summary": "Associar Diagnostico",
        "operationId": "associar_diagnostico_api_v1_jornadas__journey_id__diagnostico_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssociarDiagnosticoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JornadaPublica"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jornadas/{journey_id}/handoff": {
      "post": {
        "tags": [
          "Jornadas an�nimas"
        ],
        "summary": "Criar Handoff",
        "operationId": "criar_handoff_api_v1_jornadas__journey_id__handoff_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarHandoffRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jornadas/handoff/trocar": {
      "post": {
        "tags": [
          "Jornadas an�nimas"
        ],
        "summary": "Trocar Handoff",
        "operationId": "trocar_handoff_api_v1_jornadas_handoff_trocar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrocarHandoffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssociarDiagnosticoRequest": {
        "properties": {
          "diagnostico_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "title": "Diagnostico Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "diagnostico_id"
        ],
        "title": "AssociarDiagnosticoRequest"
      },
...
      "CriarHandoffRequest": {
        "properties": {
          "destino": {
            "type": "string",
            "title": "Destino"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "destino"
        ],
        "title": "CriarHandoffRequest"
      },
      "CriarJornadaRequest": {
        "properties": {
          "produto": {
            "$ref": "#/components/schemas/ProdutoJornada"
          },
          "mercados_basico": {
            "items": {
              "$ref": "#/components/schemas/MercadoJornada"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Mercados Basico"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "produto",
          "mercados_basico"
        ],
        "title": "CriarJornadaRequest"
      },
      "CriarJornadaResponse": {
        "properties": {
          "journey_id": {
            "type": "string",
            "title": "Journey Id"
          },
          "resume_token": {
            "type": "string",
            "title": "Resume Token"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "expira_em": {
            "type": "string",
            "format": "date-time",
            "title": "Expira Em"
          }
        },
        "type": "object",
        "required": [
          "journey_id",
          "resume_token",
          "status",
          "expira_em"
        ],
        "title": "CriarJornadaResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JornadaPublica": {
        "properties": {
          "journey_id": {
            "type": "string",
            "title": "Journey Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "produto": {
            "additionalProperties": true,
            "type": "object",
            "title": "Produto"
          },
          "mercados_basico": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Mercados Basico"
          },
          "mercados_escolhidos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mercados Escolhidos"
          },
          "diagnostico_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnostico Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "ultimo_acesso_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultimo Acesso Em"
          },
          "expira_em": {
            "type": "string",
            "format": "date-time",
            "title": "Expira Em"
          }
        },
        "type": "object",
        "required": [
          "journey_id",
          "status",
          "produto",
          "mercados_basico",
          "mercados_escolhidos",
          "diagnostico_id",
          "criado_em",
          "atualizado_em",
          "ultimo_acesso_em",
          "expira_em"
        ],
        "title": "JornadaPublica"
      },
...
      "MercadoJornada": {
        "properties": {
          "iso3": {
            "type": "string",
            "title": "Iso3"
          },
          "nome": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Nome"
          },
          "ranking_personalizado": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Ranking Personalizado"
          },
          "ranking_global_no_hs6": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranking Global No Hs6"
          },
          "score_exportai": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score Exportai"
          },
          "faixa_confianca": {
            "type": "string",
            "title": "Faixa Confianca"
          },
          "tipo_oportunidade": {
            "type": "string",
            "title": "Tipo Oportunidade"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "iso3",
          "nome",
          "ranking_personalizado",
          "score_exportai",
          "faixa_confianca",
          "tipo_oportunidade"
        ],
        "title": "MercadoJornada"
      },
...
      "PrepararDiagnosticoResponse": {
        "properties": {
          "journey_id": {
            "type": "string",
            "title": "Journey Id"
          },
          "diagnostico_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Diagnostico Payload"
          },
          "status": {
            "type": "string",
            "const": "PRONTA_PARA_DIAGNOSTICO",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "journey_id",
          "diagnostico_payload",
          "status"
        ],
        "title": "PrepararDiagnosticoResponse"
      },
...
      "ProdutoJornada": {
        "properties": {
          "ncm": {
            "type": "string",
            "title": "Ncm"
          },
          "hs6": {
            "type": "string",
            "title": "Hs6"
          },
          "descricao": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Descricao"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ncm",
          "hs6",
          "descricao"
        ],
        "title": "ProdutoJornada"
      },
...
      "SelecionarMercadosRequest": {
        "properties": {
          "mercados_escolhidos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Mercados Escolhidos"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mercados_escolhidos"
        ],
        "title": "SelecionarMercadosRequest"
      },
...
      "TrocarHandoffRequest": {
        "properties": {
          "handoff_code": {
            "type": "string",
            "title": "Handoff Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "handoff_code"
        ],
        "title": "TrocarHandoffRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}